Refactored NavigationBar
