+++ app/src/components/NavigationBar/NavigationBar.tsx
@@ -1,8 +1,6 @@
 import { Link } from 'react-router-dom';
 import {
-  IconButton,
   Button,
-  Tooltip,
   Typography,
   Box,
   AppBar,
@@ -11,10 +9,8 @@ import {
   styled,
   alpha,
   InputBase,
-  Avatar,
-  Menu,
 } from '@mui/material';
-import { useAuth } from '../hooks/FBAuthProvider';
+import { NavAvatar } from './NavAvatar';
 
 const Search = styled('div')(({ theme }) => ({
   position: 'relative',
@@ -48,59 +44,6 @@ const StyledInputBase = styled(InputBase)(({ theme }) => ({
     },
   },
 }));
-
-function AccountStatus() {
-  const auth = useAuth();
-  if (auth.user)
-    return (
-      <Box sx={{ flexGrow: 0 }}>
-        <Tooltip title="Open account settings">
-          <IconButton sx={{ p: 0 }}>
-            <Avatar alt="Z" src="../assets/images/bok.png" />
-          </IconButton>
-        </Tooltip>
-        <Menu
-          sx={{ mt: '45px' }}
-          id="menu-appbar"
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          keepMounted
-          transformOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-          open={false}
-        />
-      </Box>
-    );
-
-  return (
-    <Box sx={{ flexGrow: 0 }}>
-      {/* REGISTRATIONS BUTTON */}
-      <Button
-        component={Link}
-        to="/signup"
-        color="secondary"
-        variant="outlined"
-        size="large"
-      >
-        Registrera
-      </Button>
-      {/* LOGIN BUTTON */}
-      <Button
-        component={Link}
-        to="/login"
-        color="secondary"
-        variant="text"
-        size="large"
-      >
-        Logga in
-      </Button>
-    </Box>
-  );
-}
 
 function NavigationBar() {
   return (
@@ -152,7 +95,7 @@ function NavigationBar() {
                 />
               </Search>
             </Box>
-            <AccountStatus />
+            <NavAvatar />
           </Toolbar>
         </Container>
       </AppBar>
